Post Misc. Production Backdating issue #34641
[34641] Add config re:required trans date entry

This implements the configuration point for the optional feature allowing the
new Post Misc. Production transaction date field to require the user enter a
value rather than defaulting a value.  By default, this value is false/unchecked
meaning that the new transaction date field is defaulted to the workstation's
current date (mimicking the behavior prior to this feature existing).

[34641] Add trans date field handling

This adds handling for the new transaction date field.  Specifically we manage
the permission checks and whether or not we should default the date depending
on the AlterTransactionDates permission and the RequireMiscProdDataEntry
configuration point.  We perform a sanity check on the date prior to allowing
the user to complete a post operation.

[34641] Add transaction date field to UI

This commit adds a new transaction date field to the user interface in support
of backdating misc. production postings.

[34641] Update queries to pass transaction date

In this commit we remove the hard coded references to "CURRENT_DATE" and "now()"
where they establish the transaction date.  These date should now honor the new
transaction date field from the form.